feat: Move `ConditionalAuthComponent`
