+++ crates/miden-lib/src/testing/account_component/mod.rs
@@ -1,2 +1,5 @@
 mod incr_nonce;
 pub use incr_nonce::IncrNonceAuthComponent;
+
+mod conditional_auth;
+pub use conditional_auth::{ConditionalAuthComponent, ERR_WRONG_ARGS_MSG};
